requirement.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
better-profanity==0.7.0
certifi==2025.11.12
charset-normalizer==3.4.4
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
coloredlogs==15.0.1
datasets==4.4.2
diffusers==0.35.2
dill==0.4.0
easyocr==1.7.2
einops==0.8.1
exceptiongroup==1.3.1
filelock==3.20.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.4.2
ninja==1.13.0
nudenet==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.3
peft==0.18.0
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.0
pyarrow==22.0.0
pyclipper==1.4.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
sentencepiece==0.2.1
shapely==2.1.2
six==1.17.0
sympy==1.13.1
tifffile==2025.5.10
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
triton==3.1.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wcwidth==0.2.14
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
